自动删除上传文件的临时文件为final执行

+++ xuecheng-plus-media/xuecheng-plus-media-api/src/main/java/com/xuecheng/media/api/MediaFilesController.java
@@ -9,6 +9,7 @@ import com.xuecheng.media.model.po.MediaFiles;
 import com.xuecheng.media.service.MediaFileService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
@@ -25,6 +26,7 @@ import java.io.IOException;
  */
 @Api(value = "媒资文件管理接口", tags = "媒资文件管理接口")
 @RestController
+@Slf4j
 public class MediaFilesController {
 
 
@@ -43,27 +45,34 @@ public class MediaFilesController {
     @ApiOperation("上传图片")
     @RequestMapping(value = "/upload/coursefile", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
     public UploadFileResultDto upload(@RequestParam("filedata")MultipartFile multipartFile) throws IOException {
-        //准备上传参数数据
-        UploadparamsDto uploadparamsDto = new UploadparamsDto();
-        uploadparamsDto.setFilename(multipartFile.getOriginalFilename());  //原始文件名称
-        uploadparamsDto.setFileSize(multipartFile.getSize());  //文件大小
-        uploadparamsDto.setFileType("001001");
+
+        File tempFile = null;
+        try {
+            //准备上传参数数据
+            UploadparamsDto uploadparamsDto = new UploadparamsDto();
+            uploadparamsDto.setFilename(multipartFile.getOriginalFilename());  //原始文件名称
+            uploadparamsDto.setFileSize(multipartFile.getSize());  //文件大小
+            uploadparamsDto.setFileType("001001");
 
 
-        //创建临时文件
-        File tempFile = File.createTempFile("minio", "temp");
-        multipartFile.transferTo(tempFile);
+            //创建临时文件
+            tempFile = File.createTempFile("minio", "temp");
+            multipartFile.transferTo(tempFile);
 
-        //获取文件的路径
-        String localFilePath = tempFile.getAbsolutePath();
+            //获取文件的路径
+            String localFilePath = tempFile.getAbsolutePath();
 
-        //调用service上传图片
-        Long companyId = 12321441425L;
-        UploadFileResultDto uploadFileResultDto = mediaFileService.uploadFile(uploadparamsDto, companyId, localFilePath);
-        if (tempFile.exists()) {
-            tempFile.delete();
+            //调用service上传图片
+            Long companyId = 12321441425L;
+            return mediaFileService.uploadFile(uploadparamsDto, companyId, localFilePath);
+        } catch (Exception e) {
+            log.error("上传文件出现异常 ", e);
+            return null;
+        }  finally {
+            if (tempFile != null && tempFile.exists()) {
+                tempFile.delete();
+            }
         }
-        return uploadFileResultDto;
     }
 
 }
